Refactor: AdminService에서 AdminContainerService로 창고 로직 이동
