+++ src/main/java/com/acorn/api/controller/admin/AdminContainerManagementController.java
@@ -5,7 +5,6 @@ import com.acorn.api.code.response.ApiResponseBuilder;
 import com.acorn.api.code.response.ApiSuccessResponse;
 import com.acorn.api.dto.admin.ContainerManagementRequestDTO;
 import com.acorn.api.service.admin.AdminContainerService;
-import com.acorn.api.service.admin.AdminService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -18,14 +17,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class AdminContainerManagementController {
 
-    private final AdminService adminService;
     private final AdminContainerService adminContainerService;
 
     @PostMapping("/api/admin/container/reviewRequest/{containerId}")
     public ResponseEntity<ApiSuccessResponse<Object>> reviewRequest(@RequestBody ContainerManagementRequestDTO requestData) {
         log.info(" ************** [AdminContainerManagement] Review request started **************");
 
-        adminService.processReviewRequest(requestData);
+        adminContainerService.processReviewRequest(requestData);
 
         return ApiResponseBuilder.success(ApiSuccessCode.CONTAINER_REVIEW_SUCCESS);
     }
@@ -34,7 +32,7 @@ public class AdminContainerManagementController {
     public ResponseEntity<ApiSuccessResponse<Object>> approvalRequest(@RequestBody ContainerManagementRequestDTO requestData) {
         log.info(" ************** [AdminContainerManagement] Approval request started **************");
 
-        adminService.processApprovalRequest(requestData);
+        adminContainerService.processApprovalRequest(requestData);
 
         return ApiResponseBuilder.success(ApiSuccessCode.CONTAINER_APPROVAL_SUCCESS);
     }
@@ -43,7 +41,7 @@ public class AdminContainerManagementController {
     public ResponseEntity<ApiSuccessResponse<Object>> rejectRequest(@RequestBody ContainerManagementRequestDTO requestData) {
         log.info(" ************** [AdminContainerManagement] Reject request started **************");
 
-        adminService.processRejectRequest(requestData);
+        adminContainerService.processRejectRequest(requestData);
 
         return ApiResponseBuilder.success(ApiSuccessCode.CONTAINER_REJECT_SUCCESS);
     }
@@ -61,7 +59,7 @@ public class AdminContainerManagementController {
     public ResponseEntity<ApiSuccessResponse<Object>> cancelApproval(@RequestBody ContainerManagementRequestDTO requestData) {
         log.info(" ************** [AdminContainerManagement] Approval Cancel request started **************");
 
-        adminService.processCancelApproval(requestData);
+        adminContainerService.processCancelApproval(requestData);
 
         return ApiResponseBuilder.success(ApiSuccessCode.CONTAINER_REVIEW_SUCCESS);
     }
@@ -70,7 +68,7 @@ public class AdminContainerManagementController {
     public ResponseEntity<ApiSuccessResponse<Object>> cancelReject(@RequestBody ContainerManagementRequestDTO requestData) {
         log.info(" ************** [AdminContainerManagement] Reject Cancel request started **************");
 
-        adminService.processCancelReject(requestData);
+        adminContainerService.processCancelReject(requestData);
 
         return ApiResponseBuilder.success(ApiSuccessCode.CONTAINER_REVIEW_SUCCESS);
     }
