+++ src/main/java/com/acorn/api/controller/admin/AdminPageController.java
@@ -3,6 +3,7 @@ package com.acorn.api.controller.admin;
 import com.acorn.api.dto.admin.AdminContainerDetailResponseDTO;
 import com.acorn.api.dto.admin.AdminResponseDTO;
 import com.acorn.api.dto.container.ContainerListDTO;
+import com.acorn.api.service.admin.AdminContainerService;
 import com.acorn.api.service.admin.AdminService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -17,6 +18,7 @@ import java.util.List;
 public class AdminPageController {
 
     private final AdminService adminService;
+    private final AdminContainerService adminContainerService;
 
     @GetMapping("/admin/login")
     public String adminLoginPage() {
@@ -32,7 +34,7 @@ public class AdminPageController {
 
     @GetMapping("/admin/container/list")
     public String containersListPage(ContainerListDTO listData, Model model) {
-        List<ContainerListDTO> containerListData = adminService.getContainerList(listData);
+        List<ContainerListDTO> containerListData = adminContainerService.getContainerList(listData);
         model.addAttribute("containerList", containerListData);
         model.addAttribute("request", listData);
         return "admin/admin-container-list";
@@ -40,7 +42,7 @@ public class AdminPageController {
 
     @GetMapping("/admin/container/detail/{containerId}")
     public String containerDetailPage(@PathVariable("containerId") Integer containerId, Model model) {
-        AdminContainerDetailResponseDTO containerDetailData = adminService.getContainerData(containerId);
+        AdminContainerDetailResponseDTO containerDetailData = adminContainerService.getContainerData(containerId);
         model.addAttribute("containerDetail", containerDetailData);
         return "admin/admin-container-detail";
     }
